feat: UserController에 회원 탈퇴 시 로그인 여부 확인하는 로직 추가

+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -93,7 +93,11 @@ public class UserController {
     }
 
     @DeleteMapping("/mypage/edit/{userId}")
-    public String deleteUser(@PathVariable long userId) {
+    public String deleteUser(@PathVariable long userId, HttpSession httpSession) {
+        if (httpSession.getAttribute("name") == null) {
+            throw new UnknownHostAccessException();
+        }
+
         userService.deleteUser(userId);
         return "redirect:/logout";
     }
